Namespace MigrationSpecHelper to Spec::Support::MigrationHelper
